# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sneaker Marketplace API",
    "description": "Peer-to-peer sneaker buying and selling platform. Use the Authorize button to enter your JWT token (format: `Bearer <token>`).",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "/api", "description": "API base path" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Sneaker": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer" },
          "sellerId":   { "type": "integer" },
          "categoryId": { "type": "integer" },
          "brand":      { "type": "string" },
          "model":      { "type": "string" },
          "size":       { "type": "number" },
          "gender":     { "type": "string", "enum": ["mens","womens","unisex"] },
          "price":      { "type": "number" },
          "condition":  { "type": "string", "enum": ["new","used"] },
          "status":     { "type": "string", "enum": ["available","sold"] },
          "createdAt":  { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/sneakers": {
      "post": {
        "tags": ["Sneakers"],
        "summary": "Create a new sneaker listing",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["brand","model","size","gender","price","condition","categoryId"],
                "properties": {
                  "brand":      { "type": "string", "example": "Nike" },
                  "model":      { "type": "string", "example": "Air Jordan 1 Retro High OG" },
                  "size":       { "type": "number", "example": 10.5 },
                  "gender":     { "type": "string", "enum": ["mens","womens","unisex"], "example": "mens" },
                  "price":      { "type": "number", "example": 350.00 },
                  "condition":  { "type": "string", "enum": ["new","used"], "example": "new" },
                  "categoryId": { "type": "integer", "example": 4 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Sneaker created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Sneaker" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden: Not a seller" },
          "404": { "description": "Category not found" },
          "409": { "description": "Duplicate listing" }
        }
      },
      "get": {
        "tags": ["Sneakers"],
        "summary": "Get all sneaker listings",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "brand",      "in": "query", "schema": { "type": "string" } },
          { "name": "size",       "in": "query", "schema": { "type": "number" } },
          { "name": "categoryId", "in": "query", "schema": { "type": "integer" } },
          { "name": "maxPrice",   "in": "query", "schema": { "type": "number" } },
          { "name": "gender",     "in": "query", "schema": { "type": "string", "enum": ["mens","womens","unisex"] } }
        ],
        "responses": {
          "200": { "description": "List of sneakers", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Sneaker" } } } } },
          "400": { "description": "Invalid query parameter" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/sneakers/{id}": {
      "get": {
        "tags": ["Sneakers"],
        "summary": "Get a sneaker by ID",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Sneaker found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Sneaker" } } } },
          "400": { "description": "Invalid ID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Sneaker not found" }
        }
      },
      "put": {
        "tags": ["Sneakers"],
        "summary": "Update a sneaker listing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "brand":      { "type": "string" },
                  "model":      { "type": "string" },
                  "size":       { "type": "number" },
                  "gender":     { "type": "string", "enum": ["mens","womens","unisex"] },
                  "price":      { "type": "number" },
                  "condition":  { "type": "string", "enum": ["new","used"] },
                  "categoryId": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Sneaker updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Sneaker" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Sneaker not found" },
          "409": { "description": "Sneaker already sold" }
        }
      },
      "delete": {
        "tags": ["Sneakers"],
        "summary": "Delete a sneaker listing",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 1 }],
        "responses": {
          "200": { "description": "Sneaker deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Sneaker" } } } },
          "400": { "description": "Bad Request" },
          "401": { "description": "Unauthorized" },
          "403": { "description": "Forbidden" },
          "404": { "description": "Sneaker not found" },
          "409": { "description": "Has active order" }
        }
      }
    },
...
  }
}